improve location command
